requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
python-dotenv==0.20.0
    # via -r requirements.in
twitter==1.19.3
    # via -r requirements.in
